+++ src/main/java/net/wurstclient/clickgui/Component.java
@@ -16,7 +16,7 @@ public abstract class Component
 	
 	private Window parent;
 	
-	public void handleMouseClick(int mouseX, int mouseY, int mouseButton)
+	public void handleMouseClick(double mouseX, double mouseY, int mouseButton)
 	{
 		
 	}
